Prometheus datasource placeholder && small refactoring
