+++ src/index.ts
@@ -1,2 +1,3 @@
-export { Metric, Datasource } from './metrics';
+export { Metric } from './metrics/metrics_factory';
+export { Datasource } from './metrics/metric'
 export { queryByMetric } from './grafana_service';
