--- src/metrics/index.ts
@@ -1,2 +1,0 @@
-export { Metric } from './metrics_factory'
-export { Datasource } from './metric'
