+++ src/metrics/metric.ts
@@ -2,9 +2,9 @@ export declare type Datasource = {
   url: string;
   type: string;
   params: {
-      db: string;
-      q: string;
-      epoch: string;
+    db: string;
+    q: string;
+    epoch: string;
   };
 };
 
@@ -12,12 +12,12 @@ export type MetricId = string;
 
 export abstract class AbsractMetric {
 
-protected datasource: Datasource;
-protected targets: any[];
-protected id?: MetricId;
+  protected datasource: Datasource;
+  protected targets: any[];
+  protected id?: MetricId;
 
-constructor(datasource: Datasource, targets: any[], id?: MetricId) {}
+  constructor(datasource: Datasource, targets: any[], id?: MetricId) {}
 
-abstract getQuery(from: number, to: number, limit: number, offset: number): string;
+  abstract getQuery(from: number, to: number, limit: number, offset: number): string;
 
 }
